requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.11.1
anyio==3.7.0
async-timeout==4.0.2
asyncpg==0.27.0
attrs==23.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
ecdsa==0.18.0
email-validator==2.0.0.post2
fastapi==0.99.1
frozenlist==1.3.3
greenlet==2.0.2
h11==0.14.0
httptools==0.5.0
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.3
multidict==6.0.4
pyasn1==0.5.0
pydantic==1.10.10
python-dotenv==1.0.0
python-jose==3.3.0
PyYAML==6.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.17
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.22.0
watchfiles==0.19.0
websockets==11.0.3
yarl==1.9.2
passlib==1.7.4
python-multipart==0.0.6